+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- [updated]: Export to file: create folder when the path is invalid #5
+
 ## 0.1.0
 - [feature] new command: export to file [#2](https://github.com/Benature/obsidian-share-my-plugin-list/issues/2)
 - [updated] specific funding symbol